Added authentication to deleting roasters route

+++ controllers/requestController.js
@@ -5,6 +5,23 @@ const jwt = require("jsonwebtoken");
 const mongoose = require("mongoose");
 const Request = require("../models/requestModel")
 
+const checkAuthStatus = request => {
+    if (!request.headers.authorization) {
+        return false
+    }
+    const token = request.headers.authorization.split(" ")[1]
+
+    const loggedInUser = jwt.verify(token, process.env.JWT_SECRET, (err, data) => {
+        if (err) {
+            return false
+        }
+        else {
+            return data
+        }
+    });
+    console.log(loggedInUser)
+    return loggedInUser
+}
 
 // Get all requests
 router.get("/api/requests", async function (req, res) {
@@ -49,6 +66,11 @@ router.post("/api/requests", async function (req, res) {
 
 // Delete a request
 router.delete("/api/requests/:id", async function(req, res){
+    const loggedInUser = checkAuthStatus(req);
+    if(!loggedInUser){
+        return res.status(401).send("Must be logged in")
+    }
+    console.log(loggedInUser);
     try {
         let result = await Request.deleteOne({_id: mongoose.Types.ObjectId(req.params.id)})
         res.json(result)
